新規登録/編集画面のデザイン

+++ db/migrate/20230828014138_create_cars.rb
@@ -6,7 +6,7 @@ class CreateCars < ActiveRecord::Migration[7.0]
       t.date       :model_year,     null:false
       t.integer    :mileage,        null:false
       t.integer    :prefecture_id,  null:false
-      t.integer    :price,          null:false
+      t.float    :price,          null:false
       t.string     :car_code,       null:false, unique:true
       t.references :user,           null:false, foreign_key:true
       t.date       :car_inspection, null:false
@@ -17,6 +17,7 @@ class CreateCars < ActiveRecord::Migration[7.0]
       t.references :editer
       t.integer    :tag_id
       t.timestamps
+      
     end
   end
 end
